Move TransportConnectionOptions to its own file
